Reorganizes and documents files within project to better represent their usages
