+++ Chromatic/Extensions/Array+Ref.swift
@@ -9,6 +9,8 @@
 import Foundation
 
 extension Array {
+    // TODO: Rename or document this to better indicate that it
+    // is a safe index accessor for an array.
     func ref (_ i: Int) -> Element? {
         return 0 <= i && i < count ? self[i] : nil
     }
